requirements_clean.txt
Flask==3.1.0
python-dotenv==1.0.0
Jinja2==3.1.6
Werkzeug==3.1.3
gunicorn
google-generativeai
reportlab
Pillow
requests